# platform/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Partner Platform API",
    "version": "1.0.0",
    "description": "Affiliate, rep, and commission tracking API. Authenticate with an API key via the X-API-Key header, or sign webhook requests with HMAC-SHA256 (X-Webhook-Signature + X-Webhook-Timestamp over `${timestamp}.${rawBody}`)."
  },
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "ApiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": { "code": { "type": "string" }, "message": { "type": "string" } }
          }
        }
      },
      "OrderPayload": {
        "type": "object",
        "required": ["order_id", "subtotal", "total"],
        "properties": {
          "event": { "type": "string", "enum": ["order.completed", "order.refunded", "order.partially_refunded", "order.cancelled", "chargeback.created"], "default": "order.completed" },
          "order_id": { "type": "string" },
          "customer_id": { "type": "string" },
          "customer_email": { "type": "string", "format": "email" },
          "subtotal": { "type": "number" },
          "discount": { "type": "number", "default": 0 },
          "shipping": { "type": "number", "default": 0 },
          "tax": { "type": "number", "default": 0 },
          "total": { "type": "number" },
          "currency": { "type": "string", "default": "USD" },
          "promo_code": { "type": "string", "nullable": true },
          "affiliate_id": { "type": "string", "nullable": true, "description": "Affiliate code or referral code" },
          "click_id": { "type": "string", "nullable": true, "description": "clk_… id issued by the tracking redirect" },
          "products": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["sku", "name", "quantity", "unit_price"],
              "properties": {
                "sku": { "type": "string" },
                "name": { "type": "string" },
                "quantity": { "type": "integer" },
                "unit_price": { "type": "number" }
              }
            }
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ConversionResult": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "duplicate": { "type": "boolean" },
          "conversion_id": { "type": "string" },
          "affiliate_id": { "type": "string", "nullable": true },
          "attribution": { "type": "string", "enum": ["EXPLICIT_AFFILIATE_ID", "CLICK_ID", "PROMO_CODE", "CUSTOMER_OWNER", "COOKIE", "UNATTRIBUTED"] },
          "commission": { "type": "number" },
          "overrides": { "type": "number" }
        }
      },
...
    }
  },
  "security": [{ "ApiKey": [] }],
  "paths": {
    "/conversions": {
      "post": {
        "summary": "Ingest an order / conversion webhook",
        "description": "Requires conversion:write (or a valid HMAC signature). Supports X-Idempotency-Key for exactly-once processing; duplicate order_ids are deduplicated. Refund-type events are routed to the refund processor.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderPayload" } } } },
        "responses": {
          "201": { "description": "Conversion created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConversionResult" } } } },
          "200": { "description": "Duplicate (idempotent replay)" },
          "401": { "description": "Missing/invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "422": { "description": "Validation failed" },
          "429": { "description": "Rate limited (120 req/min per key)" }
        }
      },
      "get": {
        "summary": "List conversions",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "maximum": 100 } }
        ],
        "responses": { "200": { "description": "Paginated conversions" } }
      }
    },
...
    "/clicks": {
      "get": { "summary": "List clicks (conversion:read)", "responses": { "200": { "description": "Paginated clicks" } } },
      "post": {
        "summary": "Record a click server-side (conversion:write)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["affiliate_id"], "properties": { "affiliate_id": { "type": "string" }, "landing_page": { "type": "string" }, "utm_source": { "type": "string" }, "ip": { "type": "string" }, "user_agent": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Returns click_id" } }
      }
    },
...
  }
}